+++ components/icon/react/Icon.tsx
@@ -21,20 +21,20 @@ export const compileReactIconProperties = ({
   body,
   compiledClasses,
   size,
-  strokeColor,
-  fillColor,
-  secondaryStrokeColor,
-  secondaryFillColor,
+  interactiveColorsOnGroup,
   ...attributes
 }: {
   body: string
   compiledClasses: string[]
   size: string
-  strokeColor?
-  fillColor?
-  secondaryStrokeColor?
-  secondaryFillColor?
+  interactiveColorsOnGroup?: boolean
 } & SVGPropsWithoutColorsOrSize) => {
+  Object.keys(attributes).forEach((key) => {
+    if (key.endsWith('Color')) {
+      // @ts-ignore
+      delete attributes[key]
+    }
+  })
   const componentProps = {
     width: size,
     height: size,