+++ test/vue-app/src/App.vue
@@ -2,6 +2,7 @@
 import { ref } from 'vue'
 import Checkbox from '@cypress-design/vue-checkbox'
 import Icon, { IconObjectBookCode } from '@cypress-design/vue-icon'
+import Alert from '@cypress-design/vue-alert'
 import Spinner from '@cypress-design/vue-spinner/sfc'
 
 const checked = ref(false)
@@ -11,12 +12,12 @@ const checked = ref(false)
   <div style="width: 500px; text-align: center; margin: auto">
     <Checkbox v-model="checked" label="Checkbox" id="Sample" />
     <p v-if="checked">Checked</p>
-    <IconObjectBookCode
-      size="48"
-      strokeColor="indigo-600"
-      fillColor="indigo-300"
-    />
-    <Icon name="object-book" strokeColor="jade-500" size="24" />
+    <IconObjectBookCode size="48" :strokeColor="checked ? 'indigo-600' : 'red-300'" fillColor="indigo-300"
+      hoverFillColor="red-300" />
+    <Icon name="object-book" strokeColor="jade-500" size="24" hoverStrokeColor="red-500" />
+    <Alert type="error">
+      Error
+    </Alert>
     <Spinner />
   </div>
 </template>