+++ components/Button/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cypress-design/react-button
 
+## 0.2.1
+
+### Patch Changes
+
+- [#39](https://github.com/cypress-io/cypress-design/pull/39) [`1b383f3`](https://github.com/cypress-io/cypress-design/commit/1b383f3d149948bf2cc062d3baa17d5ce032f07e) Thanks [@elevatebart](https://github.com/elevatebart)! - fix file path of sourcemaps
+
+## 0.2.0
+
+### Minor Changes
+
+- [#35](https://github.com/cypress-io/cypress-design/pull/35) [`f68c0ed`](https://github.com/cypress-io/cypress-design/commit/f68c0ed63022f367e5859a61392539f75c3452b6) Thanks [@ryanjwilke](https://github.com/ryanjwilke)! - Renamed button types and added a new variant.
+
+* [#38](https://github.com/cypress-io/cypress-design/pull/38) [`ff7a75a`](https://github.com/cypress-io/cypress-design/commit/ff7a75aa44f6452e9041b9ec8e1d9c499106fcf1) Thanks [@mapsandapps](https://github.com/mapsandapps)! - Add new button variants and update disabled styles
+
 ## 0.1.0
 
 ### Minor Changes